# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/stakeholders": {
      "get": {
        "operationId": "getStakeholders",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number (1-based)",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "description": "Items per page",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "Column to sort by. Defaults to most recently created.",
            "schema": {
              "type": "string",
              "enum": [
                "displayId",
                "fullName",
                "createdAt",
                "lastContactedAt"
              ]
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "schema": {
              "default": "asc",
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Free-text search across fullName, email, displayId and mobile",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "clientId",
            "required": false,
            "in": "query",
            "description": "Filter by client ID (exact match)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "clientIds",
            "required": false,
            "in": "query",
            "description": "Filter by multiple client IDs at once (e.g. the stakeholder enrichment workspace, scoped to a set of selected companies)",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "jobTitle",
            "required": false,
            "in": "query",
            "description": "Filter by job title (contains, case-insensitive)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "roleTypeIds",
            "required": false,
            "in": "query",
            "description": "Filter by role type ID(s) (see /stakeholder-role-types)",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "isDecisionMaker",
            "required": false,
            "in": "query",
            "description": "Filter by decision-maker flag",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated stakeholders",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedStakeholdersEntity"
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List stakeholders (paginated, filterable, sortable)",
        "tags": [
          "Stakeholders"
        ]
      },
      "post": {
        "operationId": "createStakeholder",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateStakeholderDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Stakeholder created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StakeholderEntity"
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new stakeholder",
        "tags": [
          "Stakeholders"
        ]
      }
    },
    "/stakeholders/by-display-id/{displayId}": {
      "get": {
        "operationId": "getStakeholderByDisplayId",
        "parameters": [
          {
            "name": "displayId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Stakeholder found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StakeholderEntity"
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get stakeholder by display ID (Stake-XXXX)",
        "tags": [
          "Stakeholders"
        ]
      }
    },
    "/stakeholders/{id}": {
      "get": {
        "operationId": "getStakeholder",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Stakeholder found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StakeholderEntity"
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get stakeholder by ID",
        "tags": [
          "Stakeholders"
        ]
      },
      "patch": {
        "operationId": "updateStakeholder",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateStakeholderDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Stakeholder updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StakeholderEntity"
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a stakeholder",
        "tags": [
          "Stakeholders"
        ]
      },
      "delete": {
        "operationId": "deleteStakeholder",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Stakeholder deleted"
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a stakeholder",
        "tags": [
          "Stakeholders"
        ]
      }
    },
    "/stakeholders/{id}/contact-history": {
      "post": {
        "operationId": "addStakeholderContactHistory",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateStakeholderContactHistoryDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Contact logged",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StakeholderContactHistoryEntity"
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Log a contact with a stakeholder — the calling consultant is recorded automatically",
        "tags": [
          "Stakeholders"
        ]
      }
    },
...
  },
  "info": {
    "title": "Linktal Recruitment API",
    "description": "API for Linktal Recruitment OS",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "StakeholderEntity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "displayId": {
            "type": "string",
            "example": "Stake-0001"
          },
          "clientId": {
            "type": "string"
          },
          "companyName": {
            "type": "string",
            "nullable": true,
            "description": "Resolved client company name"
          },
          "fullName": {
            "type": "string",
            "example": "Jane Doe"
          },
          "jobTitle": {
            "type": "string",
            "nullable": true
          },
          "roleTypeId": {
            "type": "string",
            "nullable": true
          },
          "roleType": {
            "type": "string",
            "nullable": true,
            "description": "Resolved role type name"
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "mobile": {
            "type": "string",
            "nullable": true
          },
          "isDecisionMaker": {
            "type": "boolean",
            "example": false
          },
          "notes": {
            "type": "string",
            "nullable": true
          },
          "lastContactedAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true,
            "description": "Latest contactedAt across this stakeholder's own contact history; null if never contacted"
          },
          "lastContactType": {
            "type": "string",
            "nullable": true,
            "description": "Contact method of the most recent contact (email, call, meeting, linkedin)"
          },
          "lastContactNotes": {
            "type": "string",
            "nullable": true,
            "description": "Notes from the most recent contact"
          },
          "lastContactedBy": {
            "type": "string",
            "nullable": true,
            "description": "Resolved name of the consultant who made the most recent contact"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "displayId",
          "clientId",
          "companyName",
          "fullName",
          "jobTitle",
          "roleTypeId",
          "roleType",
          "email",
          "mobile",
          "isDecisionMaker",
          "notes",
          "lastContactedAt",
          "lastContactType",
          "lastContactNotes",
          "lastContactedBy",
          "createdAt",
          "updatedAt"
        ]
      },
      "PaginatedStakeholdersEntity": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StakeholderEntity"
            }
          },
          "total": {
            "type": "number",
            "description": "Total rows matching the filters",
            "example": 342
          },
          "page": {
            "type": "number",
            "description": "Current page (1-based)",
            "example": 1
          },
          "pageSize": {
            "type": "number",
            "description": "Items per page",
            "example": 20
          },
          "pageCount": {
            "type": "number",
            "description": "Total number of pages",
            "example": 18
          }
        },
        "required": [
          "data",
          "total",
          "page",
          "pageSize",
          "pageCount"
        ]
      },
      "CreateStakeholderDto": {
        "type": "object",
        "properties": {
          "clientId": {
            "type": "string",
            "description": "Client ID this stakeholder belongs to"
          },
          "fullName": {
            "type": "string",
            "description": "Full name",
            "example": "Jane Doe"
          },
          "jobTitle": {
            "type": "string",
            "description": "Job title",
            "example": "Head of Talent"
          },
          "roleTypeId": {
            "type": "string",
            "description": "Role type ID (see /stakeholder-role-types)"
          },
          "email": {
            "type": "string",
            "description": "Email address",
            "example": "jane@acme.com"
          },
          "mobile": {
            "type": "string",
            "description": "Mobile number",
            "example": "+61 412 345 678"
          },
          "isDecisionMaker": {
            "type": "boolean",
            "description": "Whether this contact is a decision maker",
            "default": false
          },
          "notes": {
            "type": "string",
            "description": "Notes"
          }
        },
        "required": [
          "clientId",
          "fullName"
        ]
      },
      "UpdateStakeholderDto": {
        "type": "object",
        "properties": {
          "clientId": {
            "type": "string",
            "description": "Client ID this stakeholder belongs to"
          },
          "fullName": {
            "type": "string",
            "description": "Full name",
            "example": "Jane Doe"
          },
          "jobTitle": {
            "type": "string",
            "description": "Job title",
            "example": "Head of Talent"
          },
          "roleTypeId": {
            "type": "string",
            "description": "Role type ID (see /stakeholder-role-types)"
          },
          "email": {
            "type": "string",
            "description": "Email address",
            "example": "jane@acme.com"
          },
          "mobile": {
            "type": "string",
            "description": "Mobile number",
            "example": "+61 412 345 678"
          },
          "isDecisionMaker": {
            "type": "boolean",
            "description": "Whether this contact is a decision maker",
            "default": false
          },
          "notes": {
            "type": "string",
            "description": "Notes"
          }
        }
      },
      "CreateStakeholderContactHistoryDto": {
        "type": "object",
        "properties": {
          "contactType": {
            "type": "string",
            "description": "How this contact happened",
            "enum": [
              "call",
              "email",
              "meeting",
              "linkedin"
            ],
            "example": "call"
          },
          "notes": {
            "type": "string",
            "description": "Notes from this contact"
          },
          "contactedAt": {
            "type": "string",
            "description": "When this contact happened (ISO 8601); defaults to now if omitted — set explicitly to log a past contact",
            "example": "2026-07-16T18:58:34.123Z"
          }
        },
        "required": [
          "contactType"
        ]
      },
      "StakeholderContactHistoryEntity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "stakeholderId": {
            "type": "string"
          },
          "contactType": {
            "type": "string",
            "example": "call"
          },
          "contactedById": {
            "type": "string",
            "nullable": true,
            "description": "Consultant who made this contact"
          },
          "notes": {
            "type": "string",
            "nullable": true
          },
          "contactedAt": {
            "format": "date-time",
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "stakeholderId",
          "contactType",
          "contactedById",
          "notes",
          "contactedAt",
          "createdAt"
        ]
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 404
          },
          "code": {
            "type": "string",
            "example": "NOT_FOUND",
            "description": "Stable machine-readable code"
          },
          "message": {
            "type": "string",
            "example": "Candidate abc not found"
          },
          "details": {
            "nullable": true,
            "description": "Field-level messages (e.g. validation errors)",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "path": {
            "type": "string",
            "example": "/api/candidates/abc"
          },
          "timestamp": {
            "type": "string",
            "example": "2026-07-09T10:20:30.000Z"
          }
        },
        "required": [
          "statusCode",
          "code",
          "message",
          "details",
          "path",
          "timestamp"
        ]
      }
    }
  }
}